Add files via upload
I used yfinance to import S&P500 data and matplotlib for plotting. Cleaning the data, I removed all unnecessary information and prepped the data by shifting it left one day to create a hypothetical "tomorrow" value to predict. I only used data from 1990 onward, as to not include irrelevant and unnecessary data from before then. In the "Target" column I created, a value of 1 is shown if the price increased from the previous day and a 0 if vice versa. Now that the data is cleaned and prepped, we can begin developing our ML algorithm to predict the next day's price.